Stream all along the way
