+++ proxy/stream.go
@@ -5,28 +5,7 @@ import (
 	"net/http"
 	"context"
 	"github.com/devopsfaith/krakend/encoding"
-	"github.com/devopsfaith/krakend/logging"
 )
-
-const streamNamespace = "github.com/devopsfaith/krakend/config/stream"
-
-func StreamConfigGetter(extra config.ExtraConfig) interface{} {
-	forward := extra["Forward"].(bool)
-	return StreamExtraConfig{forward}
-}
-
-type StreamExtraConfig struct {
-	forward bool
-}
-
-
-// DefaultFactory returns a default http proxy factory with the injected logger
-func StreamDefaultFactory(logger logging.Logger) Factory {
-	return NewDefaultFactory(streamHttpProxy, logger)
-}
-
-
-var streamHttpProxy = StreamHTTPProxyFactory(NewHTTPClient)
 
 // StreamHTTPProxyFactory returns a BackendFactory. The Proxies it creates will use the received HTTPClientFactory
 func StreamHTTPProxyFactory(cf HTTPClientFactory) BackendFactory {
@@ -38,13 +17,7 @@ func StreamHTTPProxyFactory(cf HTTPClientFactory) BackendFactory {
 
 // NewStreamHTTPProxy creates a http proxy with the injected configuration, HTTPClientFactory and Decoder
 func NewStreamHTTPProxy(cfg *config.Backend, clientFactory HTTPClientFactory, decode encoding.Decoder) Proxy {
-	streamConfigGetter := config.ConfigGetters[streamNamespace]
-	streamExtraConfig := streamConfigGetter(cfg.ExtraConfig).(StreamExtraConfig)
-	if streamExtraConfig.forward {
-		return NewHTTPStreamProxyWithHTTPExecutor(cfg, DefaultHTTPRequestExecutor(clientFactory))
-	} else {
-		return NewHTTPProxyWithHTTPExecutor(cfg, DefaultHTTPRequestExecutor(clientFactory), decode)
-	}
+	return NewHTTPStreamProxyWithHTTPExecutor(cfg, DefaultHTTPRequestExecutor(clientFactory))
 }
 
 // NewHTTPStreamProxyWithHTTPExecutor creates a http proxy with the injected configuration, HTTPRequestExecutor and Decoder
@@ -75,7 +48,6 @@ func NewHTTPStreamProxyWithHTTPExecutor(cfg *config.Backend, requestExecutor HTT
 		}
 
 		w := NewReadCloserWrapper(ctx, resp.Body)
-
 		metadata := make(map[string]string)
 
 		headers := resp.Header
